LPD-17367 Move to folder for client extension type.
