+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/js/components/global-js/ScriptElementAttributesFormField.tsx
