+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/js/index.ts
@@ -5,4 +5,4 @@
 
 export {default as ModalImportClientExtensionEntry} from './components/ModalImportClientExtensionEntry';
 export {default as remoteProtocol} from './remote_protocol/bridge';
-export {default as ScriptElementAttributesFormField} from './components/ScriptElementAttributesFormField';
+export {default as ScriptElementAttributesFormField} from './components/global-js/ScriptElementAttributesFormField';
